testes
testes

+++ app/src/main/java/api/implementation/LocalsManagement.java
@@ -24,46 +24,48 @@ import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
 public class LocalsManagement implements ILocalsManagement {
-    
-     /**
+
+    /**
      * Network graph that have information about the locals and path between them.
      */
     private IPathGameGraphADT<ILocal> pathGraph;
 
 
     /**
-     *Constructor
+     * Constructor
      */
     public LocalsManagement() {
         this.pathGraph = new PathGameGraph<>();
     }
 
-    public IPathGameGraphADT<ILocal> getPathGraph(){
+    public IPathGameGraphADT<ILocal> getPathGraph() {
         return this.pathGraph;
     }
 
     /**
      * Adds a new location to the graph
+     *
      * @param local to be added.
-     * @return string string that informs if the operation was performed successfully
+     * @return string that informs if the operation was performed successfully
      */
     @Override
-    public String addLocals (ILocal local) {
+    public String addLocals(ILocal local) {
         if (local == null) {
             throw new NullPointerException("Place cannot be null!");
         }
-         this.pathGraph.addVertex(local);
+        this.pathGraph.addVertex(local);
 
-         return "O local foi adicionado com sucesso";
+        return "O local foi adicionado com sucesso";
     }
 
     /**
      * Remove a location from the graph
+     *
      * @param local to be deleted
      * @return String that informs if the operation was performed successfully
      */
     @Override
-    public String removeLocals (ILocal local){
+    public String removeLocals(ILocal local) {
         if (local == null) {
             throw new NullPointerException("Place cannot be null!");
         }
@@ -75,11 +77,12 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Add a new interaction of a player to the connector and remove the first one from the queue if the cooldown time has already passed
+     *
      * @param connector connector where there was interaction
      * @param iteration information about the interaction to be added to the connector
      */
     @Override
-    public String addInterationConnector (IConnector connector, ConnectorPlayerInteration iteration){
+    public String addInterationConnector(IConnector connector, ConnectorPlayerInteration iteration) {
         if (connector == null) {
             throw new IllegalArgumentException("Place cannot be null!");
         }
@@ -87,18 +90,17 @@ public class LocalsManagement implements ILocalsManagement {
             throw new IllegalArgumentException("Place cannot be null!");
         }
 
-       
 
         return " Interação adicionada com sucesso com sucesso";
     }
 
     /**
      * Add new path between two points on graph.
-     * @param local1 first location
-     * @param local2 segundo local
+     *
+     * @param route to be added.
      */
     @Override
-    public String addPath(IRoute route){
+    public String addPath(IRoute route) {
         if (route.getFrom() == null || route.getTo() == null) {
             throw new NullPointerException("Place cannot be null!");
         }
@@ -111,6 +113,7 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Gets the textual listing of all Portals
+     *
      * @return string with portals listing
      */
     @Override
@@ -130,16 +133,41 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Get a textual list with the Portals that are not conquered, that is, that are not associated with any team.
+     *
      * @return string with Portals that are not conquered.
      */
     @Override
-    public String getPortalsWithoutTeamListing(){
+    public String getPortalsWithoutTeamListing() {
         String string = "Portals: {\n";
         if (this.pathGraph.getNumberOfPortals() != 0) {
             Iterator<IPortal> iteratorPortal = this.pathGraph.getPortals();
             while (iteratorPortal.hasNext()) {
                 IPortal portal = iteratorPortal.next();
-                if (portal.getPlayerTeam().equals("Neutro")){
+                if (portal.getPlayerTeam().equals("Neutro")) {
+                    string += iteratorPortal.next().toString() + "\n";
+                }
+            }
+        } else {
+            string += "There is no Portals to list!\n";
+        }
+        string += "}";
+        return string;
+    }
+
+
+    /**
+     * Get a textual list of Portals conquered by a specific player.
+     * @param player owner of the portals.
+     * @return
+     */
+    @Override
+    public String getPortalsPlayerListing(IPlayer player) {
+        String string = "Portals: {\n";
+        if (this.pathGraph.getNumberOfPortals() != 0) {
+            Iterator<IPortal> iteratorPortal = this.pathGraph.getPortals();
+            while (iteratorPortal.hasNext()) {
+                IPortal portal = iteratorPortal.next();
+                if (portal.getOwnerPlayer().equals(player)) {
                     string += iteratorPortal.next().toString() + "\n";
                 }
             }
@@ -151,42 +179,20 @@ public class LocalsManagement implements ILocalsManagement {
     }
 
     /**
-     * Get a textual list of Portals conquered by a specific player.
-     * @param player owner of the portals.
-     * @return string with all locations belonging to the player.
-     */
-    @Override
-    public String getPortalsPlayerListing(IPlayer player){
-        String string = "Portals: {\n";
-        if (this.pathGraph.getNumberOfPortals() != 0) {
-            Iterator<IPortal> iteratorPortal = this.pathGraph.getPortals();
-            while (iteratorPortal.hasNext()) {
-                IPortal portal = iteratorPortal.next();
-                if (portal.getOwnerPlayer().equals(player)){
-                        string += iteratorPortal.next().toString() + "\n";
-                }
-            }
-        } else {
-            string += "There is no Portals to list!\n";
-        }
-        string += "}";
-        return string;
-    }
-
-    /**
      * Get a textual list with the Portals conquered by a specific team.
+     *
      * @param team string with all locations belonging to the player.
      * @return string with all locations belonging to the team.
      */
     @Override
-    public String getPortalsByTeamListing(String team){
+    public String getPortalsByTeamListing(String team) {
         String string = "Portals: {\n";
         if (this.pathGraph.getNumberOfPortals() != 0) {
             Iterator<IPortal> iteratorPortal = this.pathGraph.getPortals();
             while (iteratorPortal.hasNext()) {
                 IPortal portal = iteratorPortal.next();
-                if (portal.getPlayerTeam().equals(team)){
-                        string += iteratorPortal.next().toString() + "\n";
+                if (portal.getPlayerTeam().equals(team)) {
+                    string += iteratorPortal.next().toString() + "\n";
                 }
             }
         } else {
@@ -198,10 +204,11 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Get a textual list ordered in descending order by the amount of energy the portal has.
+     *
      * @return string with the portals ordered by the amount of energy they have.
      */
     @Override
-    public String getPortalsOrderedByEnergyItHasListing(){
+    public String getPortalsOrderedByEnergyItHasListing() {
         String string = "Portals: {\n";
         List<IPortal> portals = new ArrayList<>();
         if (this.pathGraph.getNumberOfPortals() != 0) {
@@ -227,10 +234,11 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Gets the textual listing of all Connectors
+     *
      * @return string with connectors listing
      */
     @Override
-    public String getAllConnectorsListing(){
+    public String getAllConnectorsListing() {
         String string = "Connectors: {\n";
         if (this.pathGraph.getNumberOfConnectores() != 0) {
             Iterator<IConnector> iteratorConnectors = this.pathGraph.getConnectores();
@@ -246,10 +254,11 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Get a textual list of all connectors ordered in descending order by the amount of energy they have.
+     *
      * @return string with the connectors ordered according to the amount of energy they have.
      */
     @Override
-    public String getConnectorsOrderedByEnergyItHasListing(){
+    public String getConnectorsOrderedByEnergyItHasListing() {
         String string = "Connectors: {\n";
         List<IConnector> connectors = new ArrayList<>();
         if (this.pathGraph.getNumberOfPortals() != 0) {
@@ -274,9 +283,9 @@ public class LocalsManagement implements ILocalsManagement {
     }
 
 
-
     /**
      * Put all the Portals that are in the graph in a JSONArray
+     *
      * @return the JSONArray with all the locals present on the graph
      */
     @SuppressWarnings("unchecked")
@@ -292,6 +301,7 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Put all the Portals that are in the graph in a JSONArray
+     *
      * @return the JSONArray with all the locals present on the graph
      */
     @SuppressWarnings("unchecked")
@@ -306,6 +316,7 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Put all the Portals that are in the graph in a JSONArray
+     *
      * @return the JSONArray with all the locals present on the graph
      */
     @SuppressWarnings("unchecked")
@@ -322,11 +333,16 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Export all portals from a graph to a Json file
-     * @throws IOException if occurs an error trying to write the file.
+     *
      * @return A string indicating whether the operation was successful or something went wrong
+     * @throws IOException if occurs an error trying to write the file.
      */
     @Override
     public String exportPortalsToJson(String fileName) throws IOException {
+        if (fileName.trim().equals("") || Files.notExists(Paths.get(fileName))) {
+            throw new IOException("O ficheiro em que estava a tentar escrever nao existe");
+        }
+
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
         String json = gson.toJson(this.getPortalsJSONArray());
         FileWriter writer = new FileWriter(fileName);
@@ -335,17 +351,27 @@ public class LocalsManagement implements ILocalsManagement {
 
         return "O export foi feito com sucesso";
     }
-    
+
 
     /**
      * Export all Connectors from a graph to a Json file
-     * @throws IOException if occurs an error trying to write the file.
+     *
      * @return A string indicating whether the operation was successful or something went wrong.
+     * @throws IOException if occurs an error trying to write the file.
      */
     @Override
-    public String exportConnectorsToJson(String fileName) throws IOException, EmptyCollectionException {
+    public String exportConnectorsToJson(String fileName) throws IOException {
+        if (fileName.trim().equals("") || Files.notExists(Paths.get(fileName))) {
+            throw new IOException("O ficheiro em que estava a tentar escrever nao existe");
+        }
+
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        String json = gson.toJson(this.getConnectorsJSONArray());
+        String json = null;
+        try {
+            json = gson.toJson(this.getConnectorsJSONArray());
+        } catch (EmptyCollectionException e) {
+            throw new RuntimeException(e);
+        }
         FileWriter writer = new FileWriter(fileName);
         writer.write(json);
         writer.flush();
@@ -355,11 +381,16 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Export all paths from a graph to a Json file
-     * @throws IOException if occurs an error trying to write the file.
+     *
      * @return A string indicating whether the operation was successful or something went wrong
+     * @throws IOException if occurs an error trying to write the file.
      */
     @Override
-    public String exportPathsToJson(String fileName) throws IOException{
+    public String exportPathsToJson(String fileName) throws IOException {
+        if (fileName.trim().equals("") || Files.notExists(Paths.get(fileName))) {
+            throw new IOException("O ficheiro em que estava a tentar escrever nao existe");
+        }
+
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
         String json = gson.toJson(this.getRoutesJSONArray());
         FileWriter writer = new FileWriter(fileName);
@@ -371,11 +402,12 @@ public class LocalsManagement implements ILocalsManagement {
 
     /**
      * Import locations from Json file to graph
+     *
      * @param fileName fileName to use for the import
      * @return A string indicating whether the operation was successful or something went wrong
      */
     @Override
-    public String importPortalsFromJSON(String fileName) throws IOException{
+    public String importPortalsFromJSON(String fileName) throws IOException {
         if (fileName.trim().equals("") || Files.notExists(Paths.get(fileName))) {
             throw new IOException("O ficheiro em que estava a tentar escrever nao existe");
         }
@@ -391,24 +423,24 @@ public class LocalsManagement implements ILocalsManagement {
                 int amountEnergyItHas = (int) localsToCreate.get("amountEnergyItHas");
                 double longitude = (double) localsToCreate.get("longitude");
                 double latitude = (double) localsToCreate.get("latitude");
-                Coordinates coordenadas = new Coordinates(longitude,latitude);
+                Coordinates coordenadas = new Coordinates(longitude, latitude);
                 int maxEnergy = (int) localsToCreate.get("maxEnergy");
-                String teamPlayer = (String) localsToCreate.get("teamPlayer");
-                Portal portal = new Portal(maxEnergy, id, name, amountEnergyItHas,coordenadas);
+                Portal portal = new Portal(maxEnergy, id, name, amountEnergyItHas, coordenadas);
                 this.pathGraph.addVertex(portal);
             }
         } catch (ParseException e) {
-            return "Houve um problema a fazer o import dos jogadores";
+            return "Houve um problema a fazer o import dos portais";
         }
         return "O import foi feito com sucesso";
     }
 
     /**
      * Import connectors from Json file to graph
+     *
      * @param fileName fileName to use for the import
      * @return A string indicating whether the operation was successful or something went wrong
      */
-    public String importConnectorsFromJSON(String fileName) throws IOException{
+    public String importConnectorsFromJSON(String fileName) throws IOException {
         if (fileName.trim().equals("") || Files.notExists(Paths.get(fileName))) {
             throw new IOException("O ficheiro em que estava a tentar escrever nao existe");
         }
@@ -419,30 +451,51 @@ public class LocalsManagement implements ILocalsManagement {
             JSONArray jsonArray = (JSONArray) jsonObject.get("connectors");
             for (int i = 0; i < jsonArray.size(); i++) {
                 JSONObject localsToCreate = (JSONObject) jsonArray.get(i);
-                int id = (int)localsToCreate.get("id");
-                String name = (String)localsToCreate.get("name");
-                int amountEnergyItHas = (int)localsToCreate.get("amountEnergyItHas");
+                int id = (int) localsToCreate.get("id");
+                String name = (String) localsToCreate.get("name");
+                int amountEnergyItHas = (int) localsToCreate.get("amountEnergyItHas");
                 double longitude = (double) localsToCreate.get("longitude");
                 double latitude = (double) localsToCreate.get("latitude");
-                Coordinates coordenadas = new Coordinates(longitude,latitude);
-                int cooldown = (int)localsToCreate.get("cooldown");
-                Connector connector = new Connector(cooldown, id, name, amountEnergyItHas,coordenadas);
+                Coordinates coordenadas = new Coordinates(longitude, latitude);
+                int cooldown = (int) localsToCreate.get("cooldown");
+                Connector connector = new Connector(cooldown, id, name, amountEnergyItHas, coordenadas);
                 this.pathGraph.addVertex(connector);
             }
         } catch (ParseException e) {
-            return "Houve um problema a fazer o import dos jogadores";
+            return "Houve um problema a fazer o import dos connectores";
         }
         return "O import foi feito com sucesso";
     }
 
     /**
      * Import paths from a Json file into a graph
-     * @param fileName fileName fileName to use for the import
+     *
+     * @param fileName to use for the import
      * @return A string indicating whether the operation was successful or something went wrong
      */
     @Override
-    public String importPathsFromJSON(String fileName) {
+    public String importPathsFromJSON(String fileName) throws IOException {
+
+        if (fileName.trim().equals("") || Files.notExists(Paths.get(fileName))) {
+            throw new IOException("O ficheiro em que estava a tentar escrever nao existe");
+        }
+        JSONParser parser = new JSONParser();
+        try {
+            Object obj = parser.parse(new FileReader(fileName));
+            JSONObject jsonObject = (JSONObject) obj;
+            JSONArray jsonArray = (JSONArray) jsonObject.get("routes");
+            for (int i = 0; i < jsonArray.size(); i++) {
+                JSONObject routesToCreate = (JSONObject) jsonArray.get(i);
+
+                ILocal from = (ILocal) routesToCreate.get("from");
+                ILocal to = (ILocal) routesToCreate.get("to");
+
+                Route route = new Route(from, to);
+                this.addPath(route);
+            }
+        } catch (ParseException e) {
+            return "Houve um problema a fazer o import das rotas";
+        }
         return "O import foi feito com sucesso";
     }
-
 }
